Added hash methods
Added the concept of a "collection iterable" to distinguish it from "iterable".
The subtle difference is that for a collection iterable, the collection container does not matter, for an iterable it can be important.

For example a file directory can be iterable in that it gives you the files inside of it. But that does not mean you never want to do any operations on the file directory.
An arraylist of files however is a collection where only the content matters: the files.